Remove unnecessary ID from Let's Encrypt bundled cert filename
The CSR and cert_file use an ID in their filenames to ensure
they are regenerated in response to changes in relevant conditions.
However, an ID in the bundled cert filename serves no purpose.
This commit removes the ID from the bundled cert filename, offering
other programs a more stable and static path for the bundled cert.

This commit also rsyncs existing ID-in-filename bundled certs to their
non-ID counterparts. This ensures the non-ID certs exist when users
happen to run only the `wordpress` tag and not the `letsencrypt` tag.
In such a scenario, the wordpress-site.conf files would be regenerated
with non-ID filenames but such files would not exist without an rsync
effective under the same Ansible tag.